profil und user löschen

+++ php/controller/UserController.php
@@ -16,11 +16,12 @@
             if($userID >= 0){
                 $_SESSION["message"] = "login_success";
                 $_SESSION["userID"] = $userID;
-                header("Location: ../php/view/index.php");
+                error_log("user id bei login gesetzt". $_SESSION["userID"]);
+                header("Location: /php/view/index.php");
                 exit;
             }else{
                 $_SESSION["message"] = "login_failed";
-                header("Location: ../php/view/login.php");
+                header("Location: /php/view/login.php");
                 exit; 
             }
         }catch(InternalErrorException $e){
@@ -29,10 +30,10 @@
     }
 
     public function logout(){
-        //$this->checkValidUserID();
         unset($_SESSION["userID"]);
         $_SESSION["message"] = "logout_success";
-        header("Location: ../php/view/index.php");
+        header("Location: /php/view/index.php");
+        exit;
     }
 
     public function registrate(){
@@ -40,11 +41,10 @@
         try{
             $userManagement = UserManagement::getInstance();
             
-            // profilbild?
             $userManagement->saveUser($_POST["email"], $_POST["password"], $_POST["nickname"]);
 
             $_SESSION["message"] = "registration_success";
-            header("Location: ../php/view/login.php");
+            header("Location: /php/view/login.php");
             exit;
         }catch(InternalErrorException $e){
             $this->handleInternalErrorException();
@@ -69,13 +69,19 @@
     }
 
     public function deleteUser(){
-        $this->checkValidUserID();
+        if(!$this->checkValidUserID()){
+            header("Location: php/view/edit-profile.php");
+            exit;
+        }
         try{
             $userManagement = UserManagement::getInstance();
+            error_log("User löschen: ".$_SESSION["userID"]);
             $userManagement->deleteUser($_SESSION["userID"]);
+            unset($_SESSION["userID"]);
 
+        
             $_SESSION["message"] = "delete_user_success";
-            header("Location: index.php");
+            header("Location: php/view/index.php");
             exit;
         }catch(UserNotFoundException $e){
             $this->handleUserNotFoundException();
@@ -85,7 +91,10 @@
     }
 
     public function getUser(){
-        $this->checkValidUserID();
+        if(!$this->checkValidUserID()){
+            header("Location: ./index.php");
+            exit;
+        }
         try{
             $userManagement = UserManagement::getInstance();
             return $userManagement->findUser($_SESSION["userID"]);
@@ -103,11 +112,11 @@
     private function checkLoginParam(){
         if (!isset($_POST["email"]) || !isset($_POST["password"]) || !isset($_POST["submit"])) {
             $_SESSION["message"] = "missing_parameters";
-            header("Location: ../php/view/login.php");
+            header("Location: /php/view/login.php");
             exit;
         }else if(empty($_POST["email"]) || empty($_POST["password"]) ){
             $_SESSION["message"] = "missing_required_parameters";
-            header("Location: ../php/view/login.php");
+            header("Location: /php/view/login.php");
             exit;
         }
     }
@@ -117,57 +126,57 @@
      */
 
     private function checkRegistrationParam(){
-        if (!isset($_POST["email"]) || !isset($_POST["password"]) || !isset($_POST["password_confirm"]) || !isset($_POST["nickname"])|| !isset($_POST["submit"])) {
+        if (!isset($_POST["email"]) || !isset($_POST["password"]) || !isset($_POST["password_confirm"]) || !isset($_POST["nickname"]) || !isset($_POST["submit"])) {
             $_SESSION["message"] = "missing_parameters";
-            header("Location: ../php/view/registration.php");
+            header("Location: /php/view/registration.php");
             exit;
-        }else if(empty($_POST["email"]) || empty($_POST["password"]) || empty($_POST["password_confirm"]) ){
+        }else if(empty($_POST["email"]) || empty($_POST["password"]) || empty($_POST["password_confirm"]) || empty($_POST["nickname"]) ){
             $_SESSION["message"] = "missing_required_parameters";
-            header("Location: ../php/view/registration.php");
+            header("Location: /php/view/registration.php");
             exit;
         }else if ($_POST["password"] !== $_POST["password_confirm"]){
             $_SESSION["message"] = "passwords_not_equal";
-            header("Location: ../php/view/registration.php");
+            header("Location: /php/view/registration.php");
+            exit;
+        }else if (!filter_var($_POST["email"], FILTER_VALIDATE_EMAIL)){
+            $_SESSION["message"] = "invalid_email";
+            header("Location: /php/view/registration.php");
             exit;
         }
-        $this->checkValidEmail();
     }
 
     private function checkUpdateParam(){
         if (!isset($_POST["nickname"]) || !isset($_POST["profile_picture"]) || !isset($_POST["submit"])) {
             $_SESSION["message"] = "missing_parameters";
-            header("Location: ../php/view/edit-profile.php");
+            header("Location: /php/view/edit-profile.php");
             exit;
         }
-        $this->checkValidUserID();
+        else if(!$this->checkValidUserID()){
+            header("Location: /php/view/edit-profile.php");
+            exit;
+        }
     }
 
-    private function checkValidEmail(){
-        if (!filter_var($_POST["email"], FILTER_VALIDATE_EMAIL)){
-            $_SESSION["message"] = "invalid_email";
-            header("Location: ".REGISTRATION_PATH);
-            exit;
-        }
-    }
 
     private function checkValidUserID(){
-        if(!isset($_SESSION["userID"]) || empty($_SESSION["userID"])){
+        if(!isset($_SESSION["userID"])){
+            error_log("Session-Check fehlgeschlagen: userID nicht vorhanden.");
             $_SESSION["message"] = "missing_userID";
-            header("Location: ../php/view/profile.php");
-            exit;
+            return false;
         }
+        return true;
     }
 
     private function handleUserNotFoundException()
     {
         $_SESSION["message"] = "user_not_found";
-        header("Location: ../php/view/profile.php");
+        header("Location: /php/view/index.php");
         exit;
     }
     private function handleInternalErrorException()
     {
         $_SESSION["message"] = "internal_error";
-        header("Location: ../php/view/profile.php");
+        header("Location: /php/view/index.php");
         exit;
     }
     }
